some code adjustment
